langchain/requirements.txt
functions-framework==3.*
flask>=2.0.0

langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain_google_genai>=0.0.3

google-cloud-firestore>=2.11.0
google-genai>=0.7.0
google-cloud-storage

pinecone>=3.0.0

python-dotenv>=0.21.0
pydantic>=2.0.0
typing-extensions>=4.5.0
requests>=2.31.0